+++ master/internal/api_tensorboard_test.go
@@ -1,4 +1,4 @@
-package command
+package internal
 
 import (
 	"testing"
